Clear both session cookies on login redirect; refactor getUserDefaultUrlPath, isUserAdminOrUp
