+++ src/app/(main)/library/route.ts
@@ -1,16 +1,19 @@
-import { getCurrentUser, getUserDefaultUrlPath } from '@/lib/api'
+import { getCurrentUser, redirectToLogin } from '@/lib/api'
+import { isNextRedirectError } from '@/lib/nextErrors'
+import { getUserDefaultUrlPath } from '@/lib/userPermissions'
 import { redirect } from 'next/navigation'
+import { NextRequest } from 'next/server'
 
 /**
  * GET /library
  * Only serves to redirect to user default library or settings/account page
  */
-export const GET = async () => {
+export const GET = async (request: NextRequest) => {
   try {
     const currentUser = await getCurrentUser()
 
     if (!currentUser?.user) {
-      return redirect('/login')
+      return redirectToLogin(request, 'Unauthorized')
     }
 
     // Redirect to user default library or settings/account page
@@ -21,11 +24,12 @@ export const GET = async () => {
   } catch (error) {
     // Re-throw redirect errors - they are not actual errors
     // Next.js redirects throw errors with NEXT_REDIRECT in the digest
-    if (error && typeof error === 'object' && 'digest' in error && typeof error.digest === 'string' && error.digest.includes('NEXT_REDIRECT')) {
+    if (isNextRedirectError(error)) {
       throw error
     }
-    // Any other errors (including UnauthorizedError) should redirect to login
+    // Stale cookies (e.g. after DB wipe / JWT secret rotation) look valid to the proxy
+    // but fail authorize — redirectToLogin clears them so /login does not bounce back here.
     console.error('Error in library route:', error)
-    return redirect('/login')
+    return redirectToLogin(request, 'Unauthorized')
   }
 }
